requirements.txt
addict
future
lmdb
numpy
opencv-python
Pillow==9.5.0
pyyaml
requests
scikit-image
scipy
tb-nightly
torch==1.12
torchvision
tqdm
yapf
pyiqa
einops